sistema_etiquetas/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
psycopg2-binary
pymysql
Werkzeug